Convert ID types from int to long and add factory methods

+++ src/main/java/com/amalstack/notebooksfx/controller/NotebookTreeItemModel.java
@@ -6,7 +6,7 @@ import javafx.collections.ObservableList;
 public class NotebookTreeItemModel extends TreeItemModel {
     private final ObservableList<SectionTreeItemModel> sections = FXCollections.observableArrayList();
 
-    public NotebookTreeItemModel(int id, String name) {
+    public NotebookTreeItemModel(long id, String name) {
         super(id, name);
     }
